Setup Cmake, GoogleTest, and Project Environment
