+++ c_version/.gitignore
@@ -1,2 +1,2 @@
 
-main
+build
