+++ c_version/src/CMakeLists.txt
@@ -1,0 +1,2 @@
+
+add_library(math_parser SHARED lib/tokenizer.c)
