+++ c_version/test/CMakeLists.txt
@@ -1,0 +1,24 @@
+
+include(FetchContent)
+FetchContent_Declare(
+  googletest
+  URL https://github.com/google/googletest/archive/03597a01ee50ed33e9dfd640b249b4be3799d395.zip
+)
+# For Windows: Prevent overriding the parent project's compiler/linker settings
+set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
+FetchContent_MakeAvailable(googletest)
+
+enable_testing()
+
+link_libraries(
+  GTest::gtest_main
+  math_parser
+)
+
+add_executable(
+  arithmetic_test
+  arithmetic_test.cpp
+)
+
+include(GoogleTest)
+gtest_discover_tests(arithmetic_test)
